change vertex color data to usniged bytes instead of floats

+++ OpenGL/BarChart/BarChart.h
@@ -6,13 +6,14 @@
 #include <vector>
 #include <algorithm>
 #include <random>
+#include <cmath>
 
 class BarChart final
 {
 public:
 	BarChart();
 	const std::vector<GLfloat>& getRectangleVertices() const;
-	const std::vector<GLfloat>& getVertexColors() const;
+	const std::vector<GLubyte>& getVertexColors() const;
 	const std::vector<GLushort>& getIndices() const;
 	const std::vector<GLfloat>& getVerticesToSwap() const;
 	void swapVertices(const std::pair<size_t, size_t>& indexOfSwap);
@@ -20,7 +21,7 @@ public:
 
 private:
 	std::vector<GLfloat> rectangleVertices;
-	std::vector<GLfloat> vertexColors;
+	std::vector<GLubyte> vertexColors;
 	std::vector<GLushort> vertexIndices; // the indices used for the element buffer
 	std::vector<GLfloat> verticesToSwap;
